// src/domain/interfaces/repositories/_gen/SpreadsheetRepository.ts
// This file is generated by scripts/generateInterfaceRepositories.ts
import { Spreadsheet } from '../../../entities/Spreadsheet'
import { PromisedResult, UnknownRuntimeError, AlreadyExistsError, NotFoundError } from '../../../../_shared/error'

export interface SpreadsheetRepository {
  issueId: () => PromisedResult<string, UnknownRuntimeError>
  getAll: () => PromisedResult<Spreadsheet[], UnknownRuntimeError>
  getRelevant: (text: string, limit: number) => PromisedResult<Spreadsheet[], UnknownRuntimeError>
  create: (entity: Spreadsheet) => PromisedResult<Spreadsheet, UnknownRuntimeError | AlreadyExistsError>
  update: (entity: Spreadsheet) => PromisedResult<Spreadsheet, UnknownRuntimeError | NotFoundError>
  delete: (id: string) => PromisedResult<void, UnknownRuntimeError | NotFoundError>
  getById: (id: string) => PromisedResult<Spreadsheet | null, UnknownRuntimeError>
}